Moved cls Sudoku to own file #2
And add getter method for _playfield
